Split `ProductType` enum from `Common.h`
